comprende-ya-mcp/pyproject.toml
[project]
name = "comprende-ya-mcp"
version = "0.1.0"
description = "MCP server with knowledge graph for Spanish learning"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13",
    "psycopg[binary]>=3.2",
    "psycopg-pool>=3.2",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "types-pyyaml>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
ignore_missing_imports = true
